+++ wappsto/object_instantiation/instantiate.py
@@ -248,6 +248,45 @@ class Instantiator:
                         msg = "Control state {} appended to {}"
                         self.wapp_log.debug(msg.format(state, value))
 
+    def get_by_id(self, id):
+        """
+        Wappsto get by id.
+
+        Retrieves the instance of a class if its id matches the provided one
+
+        Args:
+            id: unique identifier used for searching
+
+        Returns:
+            A reference to the network/device/value/state object instance.
+
+        """
+        message = "Found instance of {} object with id: {}"
+        if self.network is not None:
+            if self.network.uuid == id:
+                self.wapp_log.debug(message.format("network", id))
+                return self.network
+
+            for device in self.network.devices:
+                if device.uuid == id:
+                    self.wapp_log.debug(message.format("device", id))
+                    return device
+
+                for value in device.values:
+                    if value.uuid == id:
+                        self.wapp_log.debug(message.format("value", id))
+                        return value
+
+                    if value.control_state is not None and value.control_state.uuid == id:
+                        self.wapp_log.debug(message.format("control state", id))
+                        return value.control_state
+
+                    if value.report_state is not None and value.report_state.uuid == id:
+                        self.wapp_log.debug(message.format("report state", id))
+                        return value.report_state
+
+        self.wapp_log.warning("Failed to find object with id: {}".format(id))
+
     def build_json(self):
         """
         Create json object.
